Generate OpenAPI validators automatically for crdexample CRDs (#6)
The way we auto-generate OpenAPI validators is as follows:

First, we add the following marker above the structures in the Go code
we want to have OpenAPI definitions for:

`// +k8s:openapi-gen=true`

Next, we use the CRD validator found here [1] to generate OpenAPI
validators for use in the CRD creation.

This is easy for the structures in types.go. However, for the Go code
which is generated from the protobuf files (e.g. the code located in
`pkg/crdexample/crdexample.proto`, the options are trickier. The
Kubernetes community around kube-openapi want to do this, as referenced
here [2], but that code isn't ready yet.

What I've done instead is write a script, `scripts/add-openapi-bits.sh`,
which will add the relevant `// +k8s:openapi-gen=true` section into the
generated Go code from the protobuf generator, and we pass that generated
file into the run of openapi-gen.

This works great, but it means that before validating our generated code
we need to do the following in sequence:

- ./scripts/update-codegen.sh
- ./scripts/add-openapi-bits.sh
- ./scripts/verify-codegen.sh

update-codegen.sh followed by add-openapi.sh ensures our generated code
matches what is in the repository itself.

Signed-off-by: Kyle Mestery <mestery@mestery.com>